fix LOGBACK-1137 and refactor RollingCalendar
https://github.com/qos-ch/logback/commit/dfa27e0c7

+++ logback-android/src/main/java/ch/qos/logback/core/CoreConstants.java
@@ -170,11 +170,11 @@ public class CoreConstants {
   public static final String VERSION_CODE_KEY = "VERSION_CODE";
 
   public static final int BYTES_PER_INT = 4;
-  public static final int MILLIS_IN_ONE_SECOND = 1000;
-  public static final int MILLIS_IN_ONE_MINUTE = MILLIS_IN_ONE_SECOND*60;
-  public static final int MILLIS_IN_ONE_HOUR = MILLIS_IN_ONE_MINUTE*60;
-  public static final int MILLIS_IN_ONE_DAY = MILLIS_IN_ONE_HOUR*24;
-  public static final int MILLIS_IN_ONE_WEEK = MILLIS_IN_ONE_DAY*7;
+  public static final long MILLIS_IN_ONE_SECOND = 1000;
+  public static final long MILLIS_IN_ONE_MINUTE = MILLIS_IN_ONE_SECOND*60;
+  public static final long MILLIS_IN_ONE_HOUR = MILLIS_IN_ONE_MINUTE*60;
+  public static final long MILLIS_IN_ONE_DAY = MILLIS_IN_ONE_HOUR*24;
+  public static final long MILLIS_IN_ONE_WEEK = MILLIS_IN_ONE_DAY*7;
 
   /**
    * The number of seconds to wait for compression jobs to finish.
